Refactored out modules and created mappers / services / models.
Created a groups module and a common module to greatly reduce the complexity of apps module.  Added in a mapping layer and model for assessments and exam to guard against the fluidly changing API.  Now when a session.id is null [Unknown] is shown.  Flattened out the sessions object to just an exams array as per what the API will be returning soon.  (The UI now determines which sessions to be filtering on).  Updated standalone data and unit tests.  Extracted out the recent assessment exam work to a new AssessmentService.  Moving forward, data calls should be coming from a service layer, and then mapping to a UI model.  This will isolate changes on the UI when API models change to just the mapping layer.