Created new user module which contains a model, mapper, service, and an authorize directive.
The authorize directive can be placed on any attribute if the user does not have the at least one of the given permissions then the given element will be removed from the DOM.

Updated standalone service to mock the api/user call.  Fixed an issue with the CachingDataService in which the subscription was “cold” in that each subscriber was re-initiating the http call (and hence not using the cached data).  The underlying structure now just saves the actual data instead of observable, and the returned observable is “shared”.