Updated calculations for aggregate info for lab vs. summative.
Added an ExamCalculator and moved the calculations here.  This will be easier to unit test.  Updated the assessment resolve to return a promise so that the route isn’t blocked when api/latestassessment doesn’t return results.

Moved enums into their own folder.  AssessmentTypeEnum now matches to the data table.