Fixed Bug 555
This broke because the translate was now taking longer to resolve now that we are merging ui / api translations.  When the primeng dropdown was being initialized, a placeholder (from translate) hadn't finished resolving so the primeng was selecting the first item in the dropdown rather than using it's placeholder.  Primeng never updated itself once the translation completed.

So now optimized the translate loader by parallelizing the api / ui translation requests and added a translate resolve so that the HomeComponent won't load until translate is complete.